+++ calamity-commands/CalamityCommands/Parser.hs
@@ -4,11 +4,16 @@ module CalamityCommands.Parser (
   Named,
   KleeneStarConcat,
   KleenePlusConcat,
+
+  -- * Parameter parsing utilities
   ParserEffs,
   runCommandParser,
   ParserState (..),
   parseMP,
+  SpannedError (..),
 ) where
+
+import CalamityCommands.ParameterInfo
 
 import Control.Lens hiding (Context)
 import Control.Monad
@@ -17,6 +22,7 @@ import Data.Char (isSpace)
 import Data.Generics.Labels ()
 import Data.Kind
 import Data.List.NonEmpty (NonEmpty (..))
+import Data.Maybe (fromMaybe)
 import Data.Semigroup
 import qualified Data.Text as S
 import qualified Data.Text.Lazy as L
@@ -38,9 +44,6 @@ import Text.Megaparsec.Char.Lexer (decimal, float, signed)
 data SpannedError = SpannedError L.Text !Int !Int
   deriving (Show, Eq, Ord)
 
-showTypeOf :: forall a. Typeable a => String
-showTypeOf = show . typeRep $ Proxy @a
-
 {- | The current state of the parser, used so that the entire remaining input is
  available.
 
@@ -56,31 +59,34 @@ data ParserState = ParserState
   }
   deriving (Show, Generic)
 
+-- |
 type ParserEffs c r =
   ( P.State ParserState
-      ': P.Error (S.Text, L.Text) -- the current parser state
-        ': P.Reader c -- (failing parser name, error reason)
+      ': P.Error (S.Text, L.Text) -- (failing parser name, error reason)
+        ': P.Reader c -- the current parser state
           ': r -- context
   )
-type ParserCtxE c r = P.Reader c ': r
 
 -- | Run a command parser, @ctx@ is the context, @t@ is the text input
 runCommandParser :: c -> L.Text -> P.Sem (ParserEffs c r) a -> P.Sem r (Either (S.Text, L.Text) a)
 runCommandParser ctx t = P.runReader ctx . P.runError . P.evalState (ParserState 0 t)
 
--- | A typeclass for things that can be parsed as parameters to commands.
---
--- Any type that is an instance of ParamerParser can be used in the type level
--- parameter @ps@ of 'CalamityCommands.Dsl.command',
--- 'CalamityCommands.CommandUtils.buildCommand', etc.
+{- | A typeclass for things that can be parsed as parameters to commands.
+
+ Any type that is an instance of ParamerParser can be used in the type level
+ parameter @ps@ of 'CalamityCommands.Dsl.command',
+ 'CalamityCommands.CommandUtils.buildCommand', etc.
+-}
 class Typeable a => ParameterParser (a :: Type) r where
   type ParserResult a
 
   type ParserResult a = a
 
-  parserName :: S.Text
-  default parserName :: S.Text
-  parserName = ":" <> S.pack (showTypeOf @a)
+  parameterInfo :: ParameterInfo
+  default parameterInfo :: ParameterInfo
+  parameterInfo = ParameterInfo Nothing (typeRep $ Proxy @a) (parameterDescription @a @r)
+
+  parameterDescription :: S.Text
 
   parse :: P.Sem (ParserEffs c r) (ParserResult a)
 
@@ -92,17 +98,32 @@ data Named (s :: Symbol) (a :: Type)
 instance (KnownSymbol s, ParameterParser a r) => ParameterParser (Named s a) r where
   type ParserResult (Named s a) = ParserResult a
 
-  parserName = (S.pack . symbolVal $ Proxy @s) <> parserName @a @r
+  parameterInfo =
+    let ParameterInfo _ type_ typeDescription = parameterInfo @a @r
+     in ParameterInfo (Just . S.pack . symbolVal $ Proxy @s) type_ typeDescription
+
+  parameterDescription = parameterDescription @a @r
 
   parse = mapE (_1 .~ parserName @(Named s a) @r) $ parse @a @r
+
+parserName :: forall a r. ParameterParser a r => S.Text
+parserName =
+  let ParameterInfo (fromMaybe "" -> name) type_ _ = parameterInfo @a @r
+   in name <> ":" <> S.pack (show type_)
 
 mapE :: P.Member (P.Error e) r => (e -> e) -> P.Sem r a -> P.Sem r a
 mapE f m = P.catch m (P.throw . f)
 
--- | Parse a paremeter using a MegaParsec parser.
---
--- On failure this constructs a nice-looking megaparsec error for the failed parameter.
-parseMP :: S.Text -> ParsecT SpannedError L.Text (P.Sem (ParserCtxE c r)) a -> P.Sem (ParserEffs c r) a
+{- | Parse a paremeter using a MegaParsec parser.
+
+ On failure this constructs a nice-looking megaparsec error for the failed parameter.
+-}
+parseMP ::
+  -- | The name of the parser
+  S.Text ->
+  -- | The megaparsec parser
+  ParsecT SpannedError L.Text (P.Sem (P.Reader c ': r)) a ->
+  P.Sem (ParserEffs c r) a
 parseMP n m = do
   s <- P.get
   res <- P.raise . P.raise $ runParserT (skipN (s ^. #off) *> trackOffsets (space *> m)) "" (s ^. #msg)
@@ -114,29 +135,37 @@ parseMP n m = do
 
 instance ParameterParser L.Text r where
   parse = parseMP (parserName @L.Text) item
+  parameterDescription = "word or quoted string"
 
 instance ParameterParser S.Text r where
   parse = parseMP (parserName @S.Text) (L.toStrict <$> item)
+  parameterDescription = "word or quoted string"
 
 instance ParameterParser Integer r where
   parse = parseMP (parserName @Integer) (signed mempty decimal)
+  parameterDescription = "number"
 
 instance ParameterParser Natural r where
   parse = parseMP (parserName @Natural) decimal
+  parameterDescription = "number"
 
 instance ParameterParser Int r where
   parse = parseMP (parserName @Int) (signed mempty decimal)
+  parameterDescription = "number"
 
 instance ParameterParser Word r where
   parse = parseMP (parserName @Word) decimal
+  parameterDescription = "number"
 
 instance ParameterParser Float r where
   parse = parseMP (parserName @Float) (signed mempty (try float <|> decimal))
+  parameterDescription = "number"
 
 instance ParameterParser a r => ParameterParser (Maybe a) r where
   type ParserResult (Maybe a) = Maybe (ParserResult a)
 
   parse = P.catch (Just <$> parse @a) (const $ pure Nothing)
+  parameterDescription = "optional " <> parameterDescription @a @r
 
 instance (ParameterParser a r, ParameterParser b r) => ParameterParser (Either a b) r where
   type ParserResult (Either a b) = Either (ParserResult a) (ParserResult b)
@@ -147,6 +176,7 @@ instance (ParameterParser a r, ParameterParser b r) => ParameterParser (Either a
       Just l' -> pure (Left l')
       Nothing ->
         Right <$> parse @b @r
+  parameterDescription = "either '" <> parameterDescription @a @r <> "' or '" <> parameterDescription @b @r <> "'"
 
 instance ParameterParser a r => ParameterParser [a] r where
   type ParserResult [a] = [ParserResult a]
@@ -159,6 +189,8 @@ instance ParameterParser a r => ParameterParser [a] r where
         Just a -> go $ l <> [a]
         Nothing -> pure l
 
+  parameterDescription = "zero or more '" <> parameterDescription @a @r <> "'"
+
 instance (ParameterParser a r, Typeable a) => ParameterParser (NonEmpty a) r where
   type ParserResult (NonEmpty a) = NonEmpty (ParserResult a)
 
@@ -166,6 +198,8 @@ instance (ParameterParser a r, Typeable a) => ParameterParser (NonEmpty a) r whe
     a <- parse @a
     as <- parse @[a]
     pure $ a :| as
+
+  parameterDescription = "one or more '" <> parameterDescription @a @r <> "'"
 
 {- | A parser that consumes zero or more of @a@ then concatenates them together.
 
@@ -177,18 +211,21 @@ instance (Monoid (ParserResult a), ParameterParser a r) => ParameterParser (Klee
   type ParserResult (KleeneStarConcat a) = ParserResult a
 
   parse = mconcat <$> parse @[a]
+  parameterDescription = "zero or more '" <> parameterDescription @a @r <> "'"
 
 instance {-# OVERLAPS #-} ParameterParser (KleeneStarConcat L.Text) r where
   type ParserResult (KleeneStarConcat L.Text) = ParserResult L.Text
 
   -- consume rest on text just takes everything remaining
   parse = parseMP (parserName @(KleeneStarConcat L.Text)) manySingle
+  parameterDescription = "the remaining input"
 
 instance {-# OVERLAPS #-} ParameterParser (KleeneStarConcat S.Text) r where
   type ParserResult (KleeneStarConcat S.Text) = ParserResult S.Text
 
   -- consume rest on text just takes everything remaining
   parse = parseMP (parserName @(KleeneStarConcat S.Text)) (L.toStrict <$> manySingle)
+  parameterDescription = "the remaining input"
 
 {- | A parser that consumes one or more of @a@ then concatenates them together.
 
@@ -200,18 +237,21 @@ instance (Semigroup (ParserResult a), ParameterParser a r) => ParameterParser (K
   type ParserResult (KleenePlusConcat a) = ParserResult a
 
   parse = sconcat <$> parse @(NonEmpty a)
+  parameterDescription = "one or more '" <> parameterDescription @a @r <> "'"
 
 instance {-# OVERLAPS #-} ParameterParser (KleenePlusConcat L.Text) r where
   type ParserResult (KleenePlusConcat L.Text) = ParserResult L.Text
 
   -- consume rest on text just takes everything remaining
   parse = parseMP (parserName @(KleenePlusConcat L.Text)) someSingle
+  parameterDescription = "the remaining input"
 
 instance {-# OVERLAPS #-} ParameterParser (KleenePlusConcat S.Text) r where
   type ParserResult (KleenePlusConcat S.Text) = ParserResult S.Text
 
   -- consume rest on text just takes everything remaining
   parse = parseMP (parserName @(KleenePlusConcat S.Text)) (L.toStrict <$> someSingle)
+  parameterDescription = "the remaining input"
 
 instance (ParameterParser a r, ParameterParser b r) => ParameterParser (a, b) r where
   type ParserResult (a, b) = (ParserResult a, ParserResult b)
@@ -220,9 +260,11 @@ instance (ParameterParser a r, ParameterParser b r) => ParameterParser (a, b) r 
     a <- parse @a
     b <- parse @b
     pure (a, b)
+  parameterDescription = "'" <> parameterDescription @a @r <> "' then '" <> parameterDescription @b @r <> "'"
 
 instance ParameterParser () r where
   parse = parseMP (parserName @()) space
+  parameterDescription = "whitespace"
 
 instance ShowErrorComponent SpannedError where
   showErrorComponent (SpannedError t _ _) = L.unpack t